feat: render rework (less code duplication)
