+++ src/marks/arc.ts
@@ -3,7 +3,7 @@ import { SceneItem } from 'vega-typings';
 import { GPUScene } from '../types/gpuscene.js';
 import { VertexBufferManager } from '../util/vertexManager.js';
 import { BufferManager } from '../util/bufferManager.js';
-import { createRenderPipeline, createUniformBindGroup, createRenderPassDescriptor } from '../util/render.js';
+import { createRenderPipeline, createUniformBindGroup, createRenderPassDescriptor } from '../util/renderer.js';
 
 import { arc } from '../path/shapes';
 import geometryForItem from '../path/geometryForItem';
