+++ src/marks/path.ts
@@ -4,7 +4,7 @@ import { SceneItem } from 'vega-typings';
 import { GPUScene } from '../types/gpuscene.js';
 import { VertexBufferManager } from '../util/vertexManager.js';
 import { BufferManager } from '../util/bufferManager.js';
-import { createRenderPipeline, createUniformBindGroup, createRenderPassDescriptor } from '../util/render.js';
+import { Renderer } from '../util/renderer.js';
 
 import geometryForItem from '../path/geometryForItem';
 import geometryForPath from '../path/geometryForPath';
@@ -35,46 +35,21 @@ function draw(device: GPUDevice, ctx: GPUCanvasContext, scene: GPUScene, vb: Bou
       ['float32x3', 'float32x4'], // position, color
       ['float32x2'] // center
     );
-    const pipeline = createRenderPipeline(drawName, device, shader, scene._format, vertextBufferManager.getBuffers());
+    const pipeline = Renderer.createRenderPipeline(drawName, device, shader, scene._format, vertextBufferManager.getBuffers());
 
     const geometryData = createGeometryData(ctx, item);
     const geometryCount = geometryData.length / vertextBufferManager.getVertexLength();
     const geometryBuffer = bufferManager.createGeometryBuffer(geometryData);
     const uniformBuffer = bufferManager.createUniformBuffer();
-    const uniformBindGroup = createUniformBindGroup(drawName, device, pipeline, uniformBuffer);
+    const uniformBindGroup = Renderer.createUniformBindGroup(drawName, device, pipeline, uniformBuffer);
     const attributes = createPosition(ctx, item);
     const instanceBuffer = bufferManager.createInstanceBuffer(attributes);
-    const frameBuffer = bufferManager.createFrameBuffer(attributes.byteLength);
-    if (geometryCount > 0) {
-      (async () => {
-        await frameBuffer.mapAsync(GPUMapMode.WRITE).then(() => {
-          const frameData = new Float32Array(frameBuffer.getMappedRange());
-          frameData.set(attributes);
 
-          const copyEncoder = device.createCommandEncoder();
-          copyEncoder.copyBufferToBuffer(
-            frameBuffer,
-            frameData.byteOffset,
-            instanceBuffer,
-            attributes.byteOffset,
-            attributes.byteLength,
-          );
-          const commandEncoder = device.createCommandEncoder();
-          const renderPassDescriptor = createRenderPassDescriptor(drawName, this.clearColor(), this.depthTexture().createView())
-          renderPassDescriptor.colorAttachments[0].view = ctx.getCurrentTexture().createView();
+    const renderPassDescriptor = Renderer.createRenderPassDescriptor(drawName, this.clearColor(), this.depthTexture().createView())
+    renderPassDescriptor.colorAttachments[0].view = ctx.getCurrentTexture().createView();
 
-          const passEncoder = commandEncoder.beginRenderPass(renderPassDescriptor);
-          passEncoder.setPipeline(pipeline);
-          passEncoder.setVertexBuffer(0, geometryBuffer);
-          passEncoder.setVertexBuffer(1, instanceBuffer);
-          passEncoder.setBindGroup(0, uniformBindGroup);
-          passEncoder.draw(geometryCount, 1, 0, 0);
-          passEncoder.end();
-          frameBuffer.unmap();
-          device.queue.submit([copyEncoder.finish(), commandEncoder.finish()]);
-        });
-      })();
-    }
+    Renderer.render2(device, pipeline, renderPassDescriptor, [geometryCount, items.length], [geometryBuffer, instanceBuffer], [uniformBindGroup]);
+
     // @ts-ignore
     ctx._tx -= item.x || 0;
     // @ts-ignore
