+++ src/marks/rule.ts
@@ -8,7 +8,7 @@ import { quadVertex } from '../util/arrays';
 import { GPUScene } from '../types/gpuscene.js';
 import { VertexBufferManager } from '../util/vertexManager.js';
 import { BufferManager } from '../util/bufferManager.js';
-import { createRenderPipeline, createUniformBindGroup, createRenderPassDescriptor } from '../util/render.js';
+import { Renderer } from '../util/renderer.js';
 
 
 const drawName = 'Rule';
@@ -32,43 +32,19 @@ function draw(device: GPUDevice, ctx: GPUCanvasContext, scene: GPUScene, vb: Bou
     // center, scale, color
     ['float32x2', 'float32x2', 'float32x4']
   );
-  const pipeline = createRenderPipeline(drawName, device, shader, scene._format, vertextBufferManager.getBuffers());
+  const pipeline = Renderer.createRenderPipeline(drawName, device, shader, scene._format, vertextBufferManager.getBuffers());
 
   const geometryBuffer = bufferManager.createGeometryBuffer(quadVertex);
   const uniformBuffer = bufferManager.createUniformBuffer();
-  const uniformBindGroup = createUniformBindGroup(drawName, device, pipeline, uniformBuffer);
+  const uniformBindGroup = Renderer.createUniformBindGroup(drawName, device, pipeline, uniformBuffer);
   const attributes = createAttributes(items);
   const instanceBuffer = bufferManager.createInstanceBuffer(attributes);
-  const frameBuffer = bufferManager.createFrameBuffer(attributes.byteLength);
-  (async () => {
-    await frameBuffer.mapAsync(GPUMapMode.WRITE).then(() => {
-      const frameData = new Float32Array(frameBuffer.getMappedRange());
-      frameData.set(attributes);
 
-      const copyEncoder = device.createCommandEncoder();
-      copyEncoder.copyBufferToBuffer(
-        frameBuffer,
-        frameData.byteOffset,
-        instanceBuffer,
-        attributes.byteOffset,
-        attributes.byteLength,
-      );
-      const commandEncoder = device.createCommandEncoder();
-      const renderPassDescriptor = createRenderPassDescriptor(drawName, this.clearColor(), this.depthTexture().createView())
-      renderPassDescriptor.colorAttachments[0].view = ctx.getCurrentTexture().createView();
+  const renderPassDescriptor = Renderer.createRenderPassDescriptor(drawName, this.clearColor(), this.depthTexture().createView())
+  renderPassDescriptor.colorAttachments[0].view = ctx.getCurrentTexture().createView();
 
-      const passEncoder = commandEncoder.beginRenderPass(renderPassDescriptor);
-      passEncoder.setPipeline(pipeline);
-      passEncoder.setVertexBuffer(0, geometryBuffer);
-      passEncoder.setVertexBuffer(1, instanceBuffer);
-      passEncoder.setBindGroup(0, uniformBindGroup);
-      // 6 because we are drawing two triangles
-      passEncoder.draw(6, items.length);
-      passEncoder.end();
-      frameBuffer.unmap();
-      device.queue.submit([copyEncoder.finish(), commandEncoder.finish()]);
-    });
-  })();
+  Renderer.render2(device, pipeline, renderPassDescriptor, [6, items.length], [geometryBuffer, instanceBuffer], [uniformBindGroup]);
+  
 }
 
 function createAttributes(items: SceneItem[]): Float32Array {
